services/api-gateway/requirements.txt
fastapi
uvicorn[standard]
requests
PyJWT[crypto]
kafka-python
prometheus_client
